Développement de l'implémentation de lof.

+++ src/main/java/lof/LOF.java
@@ -140,7 +140,7 @@ public class LOF {
 		for(int distanceAtteignabilite : distancesAtteignabilites) {
 			atteignabiliteLocale = atteignabiliteLocale + distanceAtteignabilite;
 		}
-		return k / atteignabiliteLocale;
+		return (double) k / atteignabiliteLocale;
 	}
